voice-rag/requirements.txt
streamlit
langchain_core
langchain_community
langchain_ollama
openai-whisper
